+++ autofit/graphical/__init__.py
@@ -1,12 +1,13 @@
-from . import optimise as optimize
 from . import utils
 from .declarative.abstract import PriorFactor
 from .declarative.collection import FactorGraphModel
 from .declarative.factor.analysis import AnalysisFactor
 from .declarative.factor.hierarchical import _HierarchicalFactor, HierarchicalFactor
-from .expectation_propagation import EPMeanField, EPOptimiser
+from .expectation_propagation.ep_mean_field import EPMeanField
+from .expectation_propagation.optimiser import EPOptimiser
 from .factor_graphs.factor import Factor
 from .factor_graphs.jacobians import FactorJac
 from .mean_field import FactorApproximation, MeanField
 from .optimise import OptFactor, LaplaceFactorOptimiser, lstsq_laplace_factor_approx
 from .utils import Status
+from . import optimise as optimize
