+++ autofit/graphical/expectation_propagation/ep_mean_field.py
@@ -7,10 +7,10 @@ from autoconf import cached_property
 
 from autofit.graphical.factor_graphs.factor import Factor
 from autofit.graphical.factor_graphs.graph import FactorGraph
-from autofit.graphical.mean_field import MeanField, FactorApproximation
 from autofit.graphical.utils import Status
 from autofit.mapper.variable import Variable
 from autofit.messages.abstract import AbstractMessage
+from autofit.graphical.mean_field import MeanField, FactorApproximation
 
 logger = logging.getLogger(__name__)
