+++ autofit/messages/gamma.py
@@ -3,50 +3,34 @@ from scipy import special
 
 from autoconf import cached_property
 from autofit.messages.abstract import AbstractMessage
-from autofit.graphical.utils import invpsilog
+from autofit.messages.utils import invpsilog
 
 
 class GammaMessage(AbstractMessage):
     @property
     def log_partition(self):
-        alpha, beta = GammaMessage.invert_natural_parameters(
-            self.natural_parameters
-        )
+        alpha, beta = GammaMessage.invert_natural_parameters(self.natural_parameters)
         return special.gammaln(alpha) - alpha * np.log(beta)
 
-    log_base_measure = 0.
+    log_base_measure = 0.0
     _support = ((0, np.inf),)
     _parameter_support = ((0, np.inf), (0, np.inf))
 
-    def __init__(
-            self,
-            alpha=1.,
-            beta=1.,
-            log_norm=0.,
-            id_=None
-    ):
+    def __init__(self, alpha=1.0, beta=1.0, log_norm=0.0, id_=None):
         self.alpha = alpha
         self.beta = beta
-        super().__init__(
-            alpha,
-            beta,
-            log_norm=log_norm,
-            id_=id_
-        )
+        super().__init__(alpha, beta, log_norm=log_norm, id_=id_)
 
     def value_for(self, unit: float) -> float:
         raise NotImplemented()
 
     @cached_property
     def natural_parameters(self):
-        return self.calc_natural_parameters(
-            self.alpha,
-            self.beta
-        )
+        return self.calc_natural_parameters(self.alpha, self.beta)
 
     @staticmethod
     def calc_natural_parameters(alpha, beta):
-        return np.array([alpha - 1, - beta])
+        return np.array([alpha - 1, -beta])
 
     @staticmethod
     def invert_natural_parameters(natural_parameters):
@@ -78,12 +62,7 @@ class GammaMessage(AbstractMessage):
         return np.random.gamma(a1, scale=1 / b1, size=shape)
 
     @classmethod
-    def from_mode(
-            cls,
-            mode,
-            covariance,
-            id_
-    ):
+    def from_mode(cls, mode, covariance, id_):
         m, V = cls._get_mean_variance(mode, covariance)
 
         alpha = 1 + m ** 2 * V  # match variance
@@ -96,10 +75,11 @@ class GammaMessage(AbstractMessage):
         # TODO check this is correct
         # https://arxiv.org/pdf/0911.4863.pdf
         return (
-                (P.alpha - Q.alpha) * special.psi(P.alpha)
-                - special.gammaln(P.alpha) + special.gammaln(Q.alpha)
-                + Q.alpha * (np.log(P.beta / Q.beta))
-                + P.alpha * (Q.beta / P.beta - 1)
+            (P.alpha - Q.alpha) * special.psi(P.alpha)
+            - special.gammaln(P.alpha)
+            + special.gammaln(Q.alpha)
+            + Q.alpha * (np.log(P.beta / Q.beta))
+            + P.alpha * (Q.beta / P.beta - 1)
         )
 
     def logpdf_gradient(self, x):
@@ -112,6 +92,6 @@ class GammaMessage(AbstractMessage):
         logl = self.logpdf(x)
         eta1 = self.natural_parameters[0]
         gradl = eta1 / x
-        hessl = - gradl / x
+        hessl = -gradl / x
         gradl -= self.beta
         return logl, gradl, hessl
